+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/oras-project/oras-py/tree/main) (0.0.x)
+ - TLS verify added to retry as well (0.1.27)
  - To make it available for more OCI registries, the value of config used when `manifest_config` is not specified in `client.push()` has been changed from a pure empty string to `{}` (0.1.26)
  - refactor tests using fixtures and rework pre-commit configuration (0.1.25)
  - eliminate the additional subdirectory creation while pulling an image to a custom output directory (0.1.24)